tests/requirements.txt
pillow==7.0.0
